Move plan_stubs that depend on modules other than core into their respective modules
