+++ pyproject.toml
@@ -250,7 +250,7 @@ root_package = "ophyd_async"
 name = "All runtime modules are in layers"
 type = "layers"
 containers = ["ophyd_async"]
-layers = ["plan_stubs", "fastcs", "epics | tango | sim", "core"]
+layers = ["fastcs", "epics | tango | sim", "plan_stubs", "core"]
 exhaustive = true
 exhaustive_ignores = ["testing", "_version", "__main__", "_docs_parser"]
