+++ src/ophyd_async/fastcs/panda/_plan_stubs.py
@@ -1,13 +1,13 @@
 from bluesky.utils import MsgGenerator, plan
 
 from ophyd_async.core import Settings
-from ophyd_async.fastcs import panda
+from ophyd_async.plan_stubs import apply_settings
 
-from ._settings import apply_settings
+from ._hdf_panda import HDFPanda
 
 
 @plan
-def apply_panda_settings(settings: Settings[panda.HDFPanda]) -> MsgGenerator[None]:
+def apply_panda_settings(settings: Settings[HDFPanda]) -> MsgGenerator[None]:
     """Apply given settings to a panda device."""
     units, others = settings.partition(lambda signal: signal.name.endswith("_units"))
     yield from apply_settings(units)
