+++ src/ophyd_async/plan_stubs/__init__.py
@@ -1,8 +1,6 @@
 """Plan stubs for connecting, setting up and flying devices."""
 
 from ._ensure_connected import ensure_connected
-from ._nd_attributes import setup_ndattributes, setup_ndstats_sum
-from ._panda import apply_panda_settings
 from ._settings import (
     apply_settings,
     apply_settings_if_different,
@@ -13,9 +11,6 @@ from ._settings import (
 
 __all__ = [
     "ensure_connected",
-    "setup_ndattributes",
-    "setup_ndstats_sum",
-    "apply_panda_settings",
     "apply_settings",
     "apply_settings_if_different",
     "get_current_settings",
