+++ tests/unit_tests/epics/adcore/test_writers.py
@@ -13,9 +13,9 @@ from ophyd_async.core import (
     set_mock_value,
 )
 from ophyd_async.epics import adaravis, adcore, adkinetix, adpilatus, advimba
+from ophyd_async.epics.adcore import setup_ndattributes, setup_ndstats_sum
 from ophyd_async.epics.adpilatus import PilatusReadoutTime
 from ophyd_async.epics.core import epics_signal_r
-from ophyd_async.plan_stubs import setup_ndattributes, setup_ndstats_sum
 
 DETECTOR_NAME = "test"
